requirements.txt
python-dotenv
requests
PyYAML
langchain-google-genai 
langchain-core
apscheduler